LayoutTests/imported/w3c/web-platform-tests/IndexedDB/keygenerator-inject-expected.txt


PASS Key is injected into value - single segment path
PASS Key is injected into value - multi-segment path
PASS Key is injected into value - multi-segment path, partially populated
PASS put() throws if key cannot be injected - single segment path
PASS put() throws if key cannot be injected - multi-segment path
